requirements.txt
streamlit>=1.28.0
ollama>=0.1.7
python-dotenv>=1.0.0
bcrypt>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
python-dotenv>=1.0.0
